requirements.txt
# requirements.txt
pandas
numpy
psycopg[binary]
pydantic
pydantic-settings
SQLAlchemy
python-dotenv
